% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nf_factorize_binary_column.R
\name{nf_factorize_binary_column}
\alias{nf_factorize_binary_column}
\title{Change a column that is numeric 1/0 to 'Yes'/'No', respectively}
\usage{
nf_factorize_binary_column(column, df)
}
\arguments{
\item{column}{name as a string}

\item{a}{df}
}
\value{
a df
}
\description{
Change a column that is numeric 1/0 to 'Yes'/'No', respectively
}
