% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nf_forest_plot_broom.R
\name{nf_forest_plot_broom}
\alias{nf_forest_plot_broom}
\title{Create a forest plot from the output of broom::tidy}
\usage{
nf_forest_plot_broom(df, title = NA, subtitle = NA)
}
\arguments{
\item{a}{df that broom::tidy returns}
}
\value{
a forest plot
}
\description{
Creates a forest plot from the output of broom::tidy
}
