% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nf_library.R
\name{nf_library}
\alias{nf_library}
\title{Load a vector of librarires, installing them if need be}
\usage{
nf_library(packages)
}
\arguments{
\item{packages}{a vector of package names to be loaded}
}
\value{

}
\description{
This function loads a list of librarires and installs them if necessary.
}
