% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nf_pvalue.R
\name{nf_pvalue}
\alias{nf_pvalue}
\title{Calculate the p-value for a column stratified by an outcome}
\usage{
nf_pvalue(column, outcome, df)
}
\arguments{
\item{column}{Name of the column to calculate a p-value for}

\item{outcome}{Outcome to stratify column by}
}
\value{
A cleanly formatted p-value
}
\description{
Calculate a table 1-style p-value for both continuous and categorical
outcomes
}
