% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nf_table1.R
\name{nf_table1}
\alias{nf_table1}
\title{Create a table1 and p-values}
\usage{
nf_table1(columns, outcome, df, minmax = FALSE, reset_p_value_df = FALSE)
}
\arguments{
\item{columns}{List of column names to include}

\item{outcome}{Outcome to stratify table by}
}
\value{
A table1 and p-value df
}
\description{
Create a table using the table1 package while generating p-values
}
